test-gemini.py
"""
Quick test script to verify Gemini API integration
"""

import os
from dotenv import load_dotenv
import google.generativeai as genai

# Load environment variables
load_dotenv()

# Get API key
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')
print(f"🔑 API Key loaded: {GEMINI_API_KEY[:20]}...")

# Configure Gemini
genai.configure(api_key=GEMINI_API_KEY)

print("\n🧪 Testing Gemini API Connection...\n")

try:
    # Test 1: List available models
    print("1️⃣ Listing available models...")
    models = genai.list_models()
    model_names = [m.name for m in models if 'generateContent' in m.supported_generation_methods]
    print(f"✅ Found {len(model_names)} models")
    print(f"   Available: {', '.join(model_names[:3])}...")
    
    # Test 2: Simple text generation
    print("\n2️⃣ Testing text generation (Prompt API)...")
    model = genai.GenerativeModel('gemini-2.5-flash')
    response = model.generate_content("Write a one-sentence description of AI writing assistants.")
    print(f"✅ Response: {response.text}")
    
    # Test 3: Grammar checking (Proofreader simulation)
    print("\n3️⃣ Testing grammar analysis...")
    test_text = "This are a test sentence with some error."
    prompt = f"Identify and fix grammar errors in this sentence: '{test_text}'"
    response = model.generate_content(prompt)
    print(f"✅ Grammar check result:\n   {response.text}")
    
    # Test 4: Text rewriting
    print("\n4️⃣ Testing text rewriting...")
    original = "The quick brown fox jumps over the lazy dog."
    prompt = f"Rewrite this sentence in a more formal tone: '{original}'"
    response = model.generate_content(prompt)
    print(f"✅ Rewritten text:\n   {response.text}")
    
    # Test 5: Summarization
    print("\n5️⃣ Testing summarization...")
    long_text = """
    Artificial intelligence has transformed how we write and communicate. 
    Modern AI writing assistants can help with grammar, style, tone, and even 
    content generation. They use advanced language models trained on vast 
    amounts of text to understand context and provide intelligent suggestions.
    """
    prompt = f"Summarize this text in one sentence: {long_text}"
    response = model.generate_content(prompt)
    print(f"✅ Summary:\n   {response.text}")
    
    # Test 6: Translation
    print("\n6️⃣ Testing translation...")
    english_text = "Hello, welcome to NeuroWrite!"
    prompt = f"Translate this to Spanish: '{english_text}'"
    response = model.generate_content(prompt)
    print(f"✅ Translation:\n   {response.text}")
    
    print("\n" + "="*60)
    print("🎉 ALL TESTS PASSED!")
    print("✅ Gemini API is working correctly")
    print("✅ All 6 API capabilities verified:")
    print("   1. Prompt API ✓")
    print("   2. Writer API ✓")
    print("   3. Rewriter API ✓")
    print("   4. Summarizer API ✓")
    print("   5. Translator API ✓")
    print("   6. Proofreader API ✓")
    print("="*60)
    print("\n🚀 NeuroWrite Platform is ready to launch!")
    
except Exception as e:
    print(f"\n❌ Error: {str(e)}")
    print("\nPlease check:")
    print("1. API key is correct")
    print("2. Internet connection is active")
    print("3. Gemini API is enabled for your project")
